BAQE-322: Rename tested project from dtables to rules
