+++ drools-wb-screens/drools-wb-guided-rule-editor/drools-wb-guided-rule-editor-backend/src/test/resources/org/drools/workbench/screens/guided/rule/backend/server/rules/pom.xml
@@ -3,10 +3,10 @@
     xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
   <modelVersion>4.0.0</modelVersion>
   <groupId>guvnor.feature</groupId>
-  <artifactId>dtables</artifactId>
+  <artifactId>rules</artifactId>
   <version>1.0</version>
   <packaging>kjar</packaging>
-  <name>dtables</name>
+  <name>rules</name>
   <dependencies>
     <dependency>
       <groupId>org.jboss.qa.droolsjbpm</groupId>
