+++ drools-wb-screens/drools-wb-guided-rule-editor/drools-wb-guided-rule-editor-backend/src/test/resources/org/drools/workbench/screens/guided/rule/backend/server/rules/src/main/resources/META-INF/kie-deployment-descriptor.xml
