+++ drools-wb-screens/drools-wb-guided-rule-editor/drools-wb-guided-rule-editor-backend/src/test/resources/org/drools/workbench/screens/guided/rule/backend/server/rules/src/main/resources/guvnor/feature/rules/matchAdultPeople.rdrl
@@ -1,4 +1,4 @@
-package guvnor.feature.dtables;
+package guvnor.feature.rules;
 
 import java.lang.Number;
